refactoring for OpenAI to be useable for Open Web UI
